requirements.txt
streamlit
google-generativeai
fpdf
pandas
openpyxl